+++ src/routes/web.php
@@ -125,6 +125,8 @@ Route::middleware('auth')->group(function () {
     Route::get('/my-subscriptions', [InscricoesDoAtletaController::class, 'index'])->name('subscriptions.my');
     Route::get('/minha-conta/perfil', [PerfilDoAtletaController::class, 'edit'])->name('conta.perfil');
     Route::put('/minha-conta/perfil', [PerfilDoAtletaController::class, 'update']);
+    Route::get('/minha-conta/inscricoes/{id}', [InscricoesDoAtletaController::class, 'edit'])->whereNumber('id')->name('conta.inscricao');
+    Route::put('/minha-conta/inscricoes/{id}', [InscricoesDoAtletaController::class, 'update'])->whereNumber('id');
 });
 
 Route::get('/subscribe/event/{event_id}', [SubscribeController::class, 'showSubscribeForm'])